In my last commit I'd expressed concerns about the inferred type signatures for insert, update, and filter. These were actually correct; I'd merely forgotten that Trie's are parameterized by their elements rather than sequences thereof. I've added correct signatures to these.
In my last commit I changed foldr to foldl' in singleton. This was a
mistake, as it added the reverse of the actual sequence. This has been
noted, and this function should still be sometime refactored to use
strict folds without building thunks.

Finally, I've curried singleton; now it matches the signature of update
and insert.